workers/shared-worker-sandbox.html

<!DOCTYPE html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
  const worker = new SharedWorker("support/post-origin-on-load-worker.js");
  worker.onerror = t.unreached_func("Worker construction failed");
  worker.port.onmessage = t.step_func_done(e => {
    assert_equals(e.data, "null");
  });
}, "SharedWorker is sandboxed");
</script>
</body>
